Estandarizacion de Atributos

+++ bizapps-pesaje-web/src/interfacedao/maestros/IBalanzaDAO.java
@@ -5,6 +5,6 @@ import modelo.maestros.Balanza;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IBalanzaDAO extends JpaRepository<Balanza, Integer>{
+public interface IBalanzaDAO extends JpaRepository<Balanza, Long>{
 
 }
