+++ bizapps-pesaje-web/src/interfacedao/maestros/IFleteDAO.java
@@ -5,6 +5,6 @@ import modelo.maestros.Transporte;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IFleteDAO  extends  JpaRepository<Flete, Integer>{
+public interface IFleteDAO  extends  JpaRepository<Flete, Long>{
 
 }
